edit(expose): add support for nav link prefixes
To integrate the output of this tool into my other site, I need to be
able to host its contents from a subpath. Like I did for the asset URL,
I added a new config option named `nav_link_prefix`. With that, the nav
links work when the pages are not at the root.