edit(expose, 2015, 2016): make speaker clearer
Last year we depended solely on fonts to indicate who is talking. That
worked, I think, but it certainly could be confusing. This year we
wanted to make it much clearer.

We still chose to use different fonts for each of us. But now our text
is prefixed with our initials ("TS" or "TL").