edit(2016): use `textcolor` for dark text
I forgot that Expose supports a `textcolor` setting in the frontmatter
of each Markdown file. That's what I did last year for dark text. It's
cleaner than the <span> solution I came up with this year. So I changed
the new files to use it too.

I also had to fix the links to inherit the color correctly.