add(2015): convert to a literate program
I'm a fan of literate programming. And while working on the 2016 story,
I realized this is a great program to write that way. It consists of a
lot of small files. And it's telling a story. Being able to see all of
those files in one place, within a formatted outline, is much nicer than
managing each Markdown file.